refactor: delete migration code of v1.10.1.1
